+++ test/auditing_test.exs
@@ -2,7 +2,7 @@ defmodule Commanded.Middleware.AuditingTest do
   use ExUnit.Case
 
   alias Commanded.Middleware.Auditing
-  alias Commanded.Middleware.Auditing.{CommandAudit,Repo}
+  alias Commanded.Middleware.Auditing.{CommandAudit, Repo}
   alias Commanded.Middleware.Pipeline
 
   defmodule Command do
@@ -12,20 +12,26 @@ defmodule Commanded.Middleware.AuditingTest do
   describe "before command dispatch" do
     setup [
       :execute_before_dispatch,
-      :get_audit,
+      :get_audit
     ]
 
-    test "should record command", %{pipeline: pipeline, audit: audit} do
-      assert audit != nil
-      assert is_nil audit.success
-      assert audit.occurred_at != nil
-      assert audit.occurred_at == DateTime.to_naive(pipeline.assigns.occurred_at)
+    test "should record command", %{pipeline: pipeline, audit: audit, now: now} do
+      refute is_nil(audit)
+
+      refute is_nil(audit.occurred_at)
+      assert NaiveDateTime.diff(now, audit.occurred_at) <= 1
+
       assert audit.causation_id == pipeline.causation_id
       assert audit.correlation_id == pipeline.correlation_id
       assert audit.command_uuid == pipeline.command_uuid
-      assert audit.data == "{\"age\":34,\"name\":\"Ben\",\"password\":\"[FILTERED]\",\"password_confirmation\":\"[FILTERED]\",\"secret\":\"[FILTERED]\"}"
+
+      assert audit.data ==
+               "{\"age\":34,\"name\":\"Ben\",\"password\":\"[FILTERED]\",\"password_confirmation\":\"[FILTERED]\",\"secret\":\"[FILTERED]\"}"
+
       assert audit.metadata == "{\"user\":\"user@example.com\"}"
-      assert is_nil audit.execution_duration_usecs
+
+      assert is_nil(audit.success)
+      assert is_nil(audit.execution_duration_usecs)
     end
   end
 
@@ -33,13 +39,13 @@ defmodule Commanded.Middleware.AuditingTest do
     setup [
       :execute_before_dispatch,
       :execute_after_dispatch,
-      :get_audit,
+      :get_audit
     ]
 
     test "should record success", %{audit: audit} do
       assert audit.success == true
-      assert is_nil audit.error
-      assert is_nil audit.error_reason
+      assert is_nil(audit.error)
+      assert is_nil(audit.error_reason)
       assert audit.execution_duration_usecs > 0
     end
   end
@@ -48,7 +54,7 @@ defmodule Commanded.Middleware.AuditingTest do
     setup [
       :execute_before_dispatch,
       :execute_after_failure,
-      :get_audit,
+      :get_audit
     ]
 
     test "should record failure", %{audit: audit} do
@@ -63,27 +69,36 @@ defmodule Commanded.Middleware.AuditingTest do
     setup [
       :execute_before_dispatch,
       :execute_after_failure_no_reason,
-      :get_audit,
+      :get_audit
     ]
 
     test "should record failure", %{audit: audit} do
       assert audit.success == false
       assert audit.error == ":failed"
-      assert is_nil audit.error_reason
+      assert is_nil(audit.error_reason)
       assert audit.execution_duration_usecs > 0
     end
   end
 
   defp execute_before_dispatch(_context) do
-    pipeline = Auditing.before_dispatch(%Pipeline{
-      causation_id: UUID.uuid4(),
-      correlation_id: UUID.uuid4(),
-      command: %Command{name: "Ben", age: 34, password: 1234, password_confirmation: 1234, secret: "I'm superdupersecret!"},
-      command_uuid: UUID.uuid4(),
-      metadata: %{user: "user@example.com"},
-    })
+    pipeline =
+      Auditing.before_dispatch(%Pipeline{
+        causation_id: UUID.uuid4(),
+        correlation_id: UUID.uuid4(),
+        command: %Command{
+          name: "Ben",
+          age: 34,
+          password: 1234,
+          password_confirmation: 1234,
+          secret: "I'm superdupersecret!"
+        },
+        command_uuid: UUID.uuid4(),
+        metadata: %{user: "user@example.com"}
+      })
 
-    [pipeline: pipeline]
+    now = NaiveDateTime.utc_now()
+
+    [pipeline: pipeline, now: now]
   end
 
   defp execute_after_dispatch(%{pipeline: pipeline}) do
@@ -106,7 +121,7 @@ defmodule Commanded.Middleware.AuditingTest do
     pipeline =
       pipeline
       |> Pipeline.assign(:error, :failed)
-      |> Auditing.after_failure
+      |> Auditing.after_failure()
 
     [pipeline: pipeline]
   end